Add hashes for uniqueness of runs
This doesn't protect against malicious actors, because the hashes are
selected by the frontend and not the backend, but that isn't a problem
because even if this was calculated on the backend, it would be trivial
to modify the run log in a way that keeps it as a valid run. Avoiding
that would need calculating some sort of run similarity score, and even
then it would be easily breakable.

This only avoids people double-uploading a run, which is exactly what
it should do.